Allow including namespace and schema attributes on <record>
This is useful in contexts such as OAI. OAI feeds typically wrap
individual <record> elements as children of the OAI <metadata> elements.
However, the OAI XSD for <metadata> specifies processContents="strict"
to enforce that such record elements are valid according to some XML
schema, hence the need to ensure the namespace and schema attributes are
present.